+++ karma.config.d/stubs.js
@@ -1,3 +1,3 @@
 // load stubs
-config.files.unshift("kotlin/api/SaneHandlebarsApplicationV2.js")
+config.files.unshift("kotlin/api/patches.js")
 config.files.unshift("kotlin/foundry-stubs.js")