+++ src/jsMain/kotlin/at/posselt/pfrpg/Config.kt
@@ -9,6 +9,13 @@ object Config {
         const val weather = "Weather Events"
     }
 
+    object items {
+        const val provisionsUuid = "Compendium.${moduleId}.kingmaker-tools-camping-effects.Item.UsafuPUY2soIZhC3"
+        const val specialIngredientUuid = "Compendium.pf2e.equipment-srd.Item.OCTireuX60MaPcEi"
+        const val basicIngredientUuid = "Compendium.pf2e.equipment-srd.Item.kKnMlymiqZLVEAtI"
+        const val rationUuid = "Compendium.pf2e.equipment-srd.Item.L9ZV076913otGtiB"
+    }
+
     object regions {
         const val defaultRegion = "Rostland Hinterlands"
     }