+++ src/jsMain/kotlin/at/posselt/pfrpg/app/forms/FormContext.kt
@@ -46,6 +46,7 @@ external interface FormElementContext {
     val escapeLabel: Boolean
     val templatePartial: String?
     val labelElement: String
+    val placeholder: String?
 
     // instead of going with an enum, we need to list all values
     // because Handlebars is fucking stupid and can't have == in their template language