+++ src/jsMain/kotlin/at/posselt/pfrpg/camping/CampingChat.kt
@@ -1,10 +1,14 @@
 package at.posselt.pfrpg.camping
 
+import at.posselt.pfrpg.Config
 import at.posselt.pfrpg.utils.bindChatClick
 import at.posselt.pfrpg.utils.buildPromise
 import at.posselt.pfrpg.utils.emitPfrpg2eKingdomCampingWeather
+import at.posselt.pfrpg.utils.fromUuidTypeSafe
 import at.posselt.pfrpg.utils.isFirstGM
+import at.posselt.pfrpg.utils.postChatMessage
 import com.foundryvtt.core.Game
+import com.foundryvtt.pf2e.actor.PF2ECharacter
 import org.w3c.dom.get
 
 fun bindCampingChatEventListeners(game: Game) {
@@ -12,6 +16,16 @@ fun bindCampingChatEventListeners(game: Game) {
         game.getCampingActor()?.let { actor ->
             buildPromise {
                 rollRandomEncounter(game, actor, true)
+            }
+        }
+    }
+    bindChatClick(".gain-provisions") { _, el ->
+        buildPromise {
+            val actor = el.dataset["actorUuid"]?.let { fromUuidTypeSafe<PF2ECharacter>(it) }
+            val quantity = el.dataset["quantity"]?.toInt() ?: 0
+            if (quantity > 0 && actor != null) {
+                actor.addConsumableToInventory(Config.items.provisionsUuid, quantity)
+                postChatMessage("${actor.name}: Adding $quantity provisions")
             }
         }
     }