+++ src/jsMain/kotlin/at/posselt/pfrpg/camping/Food.kt
@@ -1,5 +1,6 @@
 package at.posselt.pfrpg.camping
 
+import at.posselt.pfrpg.Config
 import at.posselt.pfrpg.divideRoundingUp
 import at.posselt.pfrpg.utils.buildUpdate
 import at.posselt.pfrpg.utils.fromUuidTypeSafe
@@ -19,10 +20,6 @@ import kotlinx.js.JsPlainObject
 import kotlin.js.unsafeCast
 import kotlin.math.max
 import kotlin.math.min
-
-private const val specialIngredientUuid = "Compendium.pf2e.equipment-srd.Item.OCTireuX60MaPcEi"
-private const val basicIngredientUuid = "Compendium.pf2e.equipment-srd.Item.kKnMlymiqZLVEAtI"
-private const val rationUuid = "Compendium.pf2e.equipment-srd.Item.L9ZV076913otGtiB"
 
 suspend fun PF2EActor.addConsumableToInventory(uuid: String, quantity: Int) {
     if (quantity > 0) {
@@ -160,9 +157,9 @@ fun List<FoodAmount>.sum() =
 
 suspend fun PF2EActor.addFoodToInventory(foodAmount: FoodAmount) = coroutineScope {
     listOf(
-        async { addConsumableToInventory(specialIngredientUuid, foodAmount.specialIngredients) },
-        async { addConsumableToInventory(basicIngredientUuid, foodAmount.basicIngredients) },
-        async { addConsumableToInventory(rationUuid, foodAmount.rations) },
+        async { addConsumableToInventory(Config.items.specialIngredientUuid, foodAmount.specialIngredients) },
+        async { addConsumableToInventory(Config.items.basicIngredientUuid, foodAmount.basicIngredients) },
+        async { addConsumableToInventory(Config.items.rationUuid, foodAmount.rations) },
     ).awaitAll()
 }
 
@@ -227,7 +224,8 @@ suspend fun reduceFoodBy(actors: List<PF2EActor>, foodAmount: FoodAmount, foodIt
 suspend fun PF2EActor.reduceFoodBy(foodAmount: FoodAmount, foodItems: FoodItems): FoodAmount = coroutineScope {
     val leftOverSpecial = async { removeConsumableFromInventory(foodItems.special.slug, foodAmount.specialIngredients) }
     val leftOverBasic = async { removeConsumableFromInventory(foodItems.basic.slug, foodAmount.basicIngredients) }
-    val leftOverRation = async { removeConsumableFromInventory(foodItems.ration.slug, foodAmount.rations) }
+    val leftOverProvisions = async { removeConsumableFromInventory(foodItems.provisions.slug, foodAmount.rations) }
+    val leftOverRation = async { removeConsumableFromInventory(foodItems.ration.slug, leftOverProvisions.await()) }
     FoodAmount(
         basicIngredients = leftOverBasic.await(),
         specialIngredients = leftOverSpecial.await(),
@@ -258,7 +256,8 @@ private fun findTotalFood(
     return FoodAmount(
         basicIngredients = actor.consumableQuantityBySlug(foodItems.basic.slug),
         specialIngredients = actor.consumableQuantityBySlug(foodItems.special.slug),
-        rations = actor.consumableQuantityBySlug(foodItems.ration.slug),
+        rations = actor.consumableQuantityBySlug(foodItems.ration.slug) +
+                actor.consumableQuantityBySlug(foodItems.provisions.slug),
     )
 }
 
@@ -266,19 +265,23 @@ data class FoodItems(
     val basic: PF2EConsumable,
     val special: PF2EConsumable,
     val ration: PF2EConsumable,
+    val provisions: PF2EConsumable,
 )
 
 suspend fun getCompendiumFoodItems() = coroutineScope {
-    val b = async { fromUuidTypeSafe<PF2EConsumable>(basicIngredientUuid) }
-    val s = async { fromUuidTypeSafe<PF2EConsumable>(specialIngredientUuid) }
-    val r = async { fromUuidTypeSafe<PF2EConsumable>(rationUuid) }
+    val b = async { fromUuidTypeSafe<PF2EConsumable>(Config.items.basicIngredientUuid) }
+    val s = async { fromUuidTypeSafe<PF2EConsumable>(Config.items.specialIngredientUuid) }
+    val r = async { fromUuidTypeSafe<PF2EConsumable>(Config.items.rationUuid) }
+    val p = async { fromUuidTypeSafe<PF2EConsumable>(Config.items.provisionsUuid) }
     val basic = b.await()
     val special = s.await()
     val ration = r.await()
+    val provisions = p.await()
     checkNotNull(basic) { "Basic Ingredient UUID changed, something is very wrong" }
     checkNotNull(special) { "Basic Ingredient UUID changed, something is very wrong" }
     checkNotNull(ration) { "Basic Ingredient UUID changed, something is very wrong" }
-    FoodItems(basic = basic, special = special, ration = ration)
+    checkNotNull(provisions) { "Basic Ingredient UUID changed, something is very wrong" }
+    FoodItems(basic = basic, special = special, ration = ration, provisions = provisions)
 }
 
 suspend fun CampingData.getFoodAmount(