+++ src/jsMain/kotlin/com/foundryvtt/core/Roll.kt
@@ -18,10 +18,14 @@ external interface RollMessageOptions {
     val create: Boolean?
 }
 
-external class Roll(formula: String) {
+open external class Roll(formula: String) {
     val formula: String
     val result: String
     val total: Int
-    fun toMessage(data: AnyObject, options: RollMessageOptions = definedExternally): Promise<ChatMessage>
+    fun toMessage(
+        data: AnyObject = definedExternally,
+        options: RollMessageOptions = definedExternally
+    ): Promise<ChatMessage>
+
     fun evaluate(context: RollContext = definedExternally): Promise<Roll>
 }